+++ src/test/java/tv/ustream/yolo/module/parser/JsonParserTest.java
@@ -1,5 +1,6 @@
 package tv.ustream.yolo.module.parser;
 
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.Assert;
@@ -99,6 +100,62 @@ public class JsonParserTest
         Assert.assertEquals(expected, actual);
     }
 
+    @Test
+    public void parserShouldAllowWithKeyFilter() throws Exception
+    {
+        Map<String, Object> expected = new HashMap<String, Object>();
+        expected.put("key1", null);
+        expected.put("key3", "x");
+
+        parser = createParserWithFilter();
+
+        Map<String, Object> actual = parser.parse(
+            "{\"key1\":null,\"key3\":\"x\"}"
+        );
+
+        Assert.assertEquals(expected, actual);
+    }
+
+    @Test
+    public void parserShouldAllowWithKeyValueFilter() throws Exception
+    {
+        Map<String, Object> expected = new HashMap<String, Object>();
+        expected.put("key2", "value2");
+        expected.put("key3", "x");
+
+        parser = createParserWithFilter();
+
+        Map<String, Object> actual = parser.parse(
+            "{\"key2\":\"value2\",\"key3\":\"x\"}"
+        );
+
+        Assert.assertEquals(expected, actual);
+    }
+
+    @Test
+    public void parserShouldDenyWhenKeyDoesNotExist() throws Exception
+    {
+        parser = createParserWithFilter();
+
+        Map<String, Object> actual = parser.parse(
+            "{\"key2\":null,\"key3\":\"y\"}"
+        );
+
+        Assert.assertNull(actual);
+    }
+
+    @Test
+    public void parserShouldDenyWhenValueIsDifferent() throws Exception
+    {
+        parser = createParserWithFilter();
+
+        Map<String, Object> actual = parser.parse(
+            "{\"key3\":\"x\",\"key4\":\"y\"}"
+        );
+
+        Assert.assertNull(actual);
+    }
+
     private IParser createParser() throws Exception
     {
         return createParser(true);
@@ -113,6 +170,26 @@ public class JsonParserTest
         config.put("class", JsonParser.class.getCanonicalName());
         config.put("processors", processors);
         config.put("flatten", flatten);
+        return new ModuleFactory().createParser("x", config);
+    }
+
+    private IParser createParserWithFilter() throws Exception
+    {
+        Map<String, Object> processors = new HashMap<String, Object>();
+        processors.put("processor1", new HashMap<String, Object>());
+
+        Map<String, Object> filter1 = new HashMap<>();
+        filter1.put("key", "key1");
+
+        Map<String, Object> filter2 = new HashMap<>();
+        filter2.put("key", "key2");
+        filter2.put("value", "value2");
+
+        Map<String, Object> config = new HashMap<String, Object>();
+        config.put("class", JsonParser.class.getCanonicalName());
+        config.put("processors", processors);
+        config.put("flatten", true);
+        config.put("filters", Arrays.asList(filter1, filter2));
         return new ModuleFactory().createParser("x", config);
     }